[JT-63] refactor: 이미지 업로드 DTO 수정
