+++ module-application/app-api/src/main/java/shop/jtoon/webtoon/request/CreateWebtoonReq.java
@@ -2,10 +2,15 @@ package shop.jtoon.webtoon.request;
 
 import java.util.Set;
 
+import org.springframework.web.multipart.MultipartFile;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import shop.jtoon.common.FileName;
+import shop.jtoon.common.ImageType;
 import shop.jtoon.dto.CreateWebtoonDto;
+import shop.jtoon.dto.UploadImageDto;
 import shop.jtoon.entity.Member;
 import shop.jtoon.entity.enums.AgeLimit;
 import shop.jtoon.entity.enums.DayOfWeek;
@@ -30,6 +35,15 @@ public record CreateWebtoonReq(
 			.genres(genres)
 			.ageLimit(ageLimit)
 			.cookieCount(cookieCount)
+			.build();
+	}
+
+	public UploadImageDto toUploadImageDto(ImageType imageType, MultipartFile image) {
+		return UploadImageDto.builder()
+			.imageType(imageType)
+			.webtoonTitle(title)
+			.fileName(FileName.forWebtoon())
+			.image(image)
 			.build();
 	}
 }
