+++ module-internal/s3-client/src/main/java/shop/jtoon/dto/UploadImageDto.java
@@ -1,4 +1,4 @@
-package shop.jtoon.request;
+package shop.jtoon.dto;
 
 import org.springframework.web.multipart.MultipartFile;
 
@@ -7,21 +7,12 @@ import shop.jtoon.common.FileName;
 import shop.jtoon.common.ImageType;
 
 @Builder
-public record UploadImageReq(
+public record UploadImageDto(
 	ImageType imageType,
 	String webtoonTitle,
 	FileName fileName,
 	MultipartFile image
 ) {
-
-	public static UploadImageReq of(ImageType imageType, String webtoonTitle, FileName fileName, MultipartFile image) {
-		return UploadImageReq.builder()
-			.imageType(imageType)
-			.webtoonTitle(webtoonTitle)
-			.fileName(fileName)
-			.image(image)
-			.build();
-	}
 
 	public String toKey() {
 		return imageType.getPath(webtoonTitle, fileName.getValue());
